libdi: implement async open customization point
Now, async_read and async_write take an arbitrary handle as input,
which is meant to result from a call to async_open(). This allows
opening/closing using purely asynchronous operations.

Note that the current implementation of IoContext in libdius performs
synchronous operations, but will use io_uring in the future.

From the sample code performing an asynchronous copy, it seems that
asynchronous resources could benefit from a way to pass multiple
senders in directly to a call to execution::with(). This would reduce
nesting and allow for opening/closing of 2 separate files to happen
concurrently.