libdi: add custom platform.h file which is user controllable
This enables different platforms to provide default implementation
for some core library concepts, such as the default lock and default
allocator. It also provides a way to access the current thread id,
which is very useful for some synchronization algorithms.

For kernel context, this will allow easily using a fallible allocator
as the default, as well as a OS provided mutex instead of std::mutex.

For single threaded enviornments, a custom lock could essentially be
a no-op, and the thread id would statically return 0.

In the future, this mechanism should be used to provide the customization
of the assert macros as well.