libdius: initial io_uring context
This is a very rough first-pass implementation, and so is not
thread-safe, or even correct. However, it is sufficent to run the
single threaded asynchronous copy program.

The implementation currently uses liburing, but in the future I
would like to implement a wrapper written in pure C++ instead.

The main idea necessary to support access from multiple threads is
to have an eventfd which the event loop waits on when submitting
entries. Then, when an operation is scheduled from a different thread,
this operation will be put into a queue and the eventfd will be notified.
This will wake up the main loop, and then it will poll the queue.

There is further work to be done on investigating how to implement timers
on top of this context (timerfd, linked OP_TIMEOUT, submit w/ timeout),
and how to handle cancellation (OP_CANCEL).