libdi: add support for null-terminating strings
Now, an encoding can make itself null-terminated, and owning strings
using this encoding will then be null-terminated. To prevent allocation
in the default case, an empty string is internally still represented
by nullptr. To get an actual null-terminated string the .c_str()
method must be called, which checks if the underlying store is empty,
and will return a pointer to a static buffer containing a zero.

Unfortunately, null-terminated strings are still relevant when interfacing
with C APIs, so adding support for them is necessary. Still, it won't be
the default for UTF-8 strings.

The transparent string encoding now is null-terminated by default, which
reflects the fact that UNIX paths are null-terminated, and the transparent
encoding is used for path objects by default.