libdius: implement async open/close for io_uring
Note that instead of taking a PathView for the path to open, the
asynchronous version takes a owned Path with is also null-terminated.
This prevents the need to copy the path when performing the open,
and also ensures that the path does not go out scope before the
operation is submitted to the kernel.