libdi+libdius: rename io read() to read_some() or read_exactly()
The underlying system call APIs for file IO return the number of
bytes read or written, which may be less than requested. In many
cases, application code wants to make sure all bytes are read or
written.

Examples include copying a file, where all bytes need to be written,
or reading from a fixed offset in a ELF file to construct the symbol
table.

Therefore, the library now makes a distinction between xxx_some()
and xxx_exactly(), and xxx_exactly() can be implemented by calling
xxx_some() in a loop. This will enable application code to ignore
the corner case where a read or write call is partial, unless it
is capable of handling it.