edit: fix broken saving of files
This removes the chdir() call at the start of edit, since it causes
the document's file path to become invalidated. This later on prevents
files from being saved.

In addition, this patch includes a work-around for some sort of GCC
regression in 12.2. Note that the code may have been relying on
non-standard behavior before, so this is not necessarily a problem
in GCC.