iris: implement cooperative scheduling of kernel tasks
This is mainly a stepping stone towards pre-emptive scheduling, but
the Scheduler::yield() function is still quite useful, especially
when implementing a kernel-space mutex.

The code to save the task state is mostly copy-pasted from the C
kernel, and so could be considerably improved. It would also be nice
to have an easy way to call C++ manged symbols from assembly, although
this would presumably require implementing C++ name mangling on the
C preprocessor.