iris: generate IRQ handlers for each specific interrupt
Since each vector number needs an installed IRQ handler, but most
of its code is redundant, each IRQ handler is generated from its
vector number and whether or not it pushes an error code onto the
stack.

Since this process is naturally repetetive, it is a prime candidate
for using the C preprocessor. Essentially, we can define a macro
which takes another macro as argument and invokes said macro on
every number from 0 to 255 inclusive. This is used to fill in the
IDT and generate interrupt handlers for each interrupt vector.

We additionally need to make use of preprocessor conditionals
to determine if the vector pushes onto the stack. This is done
using pattern matching, by appending the interrupt number into
a macro, which only expands if it is defined.

These preprocessor tricks are descripted in this blog post:
http://jhnet.co.uk/articles/cpp_magic