libdi: implement MDSpan iteration and LayoutStride
This completes the parts of MDSpan that are part of c++ 23. However,
there are still additional thinks that can be implemented, like
mdarray and submdspan.

Iterating over all elements of an MDSpan uses di::cartesian_product().
This implementation is very inefficent, especially considering that
the default layout would only need to access each underlying element
sequentially. This feature is desirable because MDSpan with a LayoutStride
is perfect for representing on-disk data structures which are forward compatible
by providing a size field. For instance, the ext2 file system can have inodes
of either 128 or 256 bytes in size. To store a span of these safely, we can use
a 1 dimensional LayoutStride with MDSpan, which is significantly safer than
performing raw pointer arithmetic.

This approach won't work if some data structure
is, for instance, 12 bytes, and then its size get extended to a non-multiple of 12
(16). Because of this, there may need to be a separate type which is a non-owning
view with byte index strides. Alternatively, a custom MDSpan accessor could be used
with an MDSpan of bytes. This accessor would cast the underlying bytes to the desired
type.