libdius: implement system calls without using libc
This patch replaces calls to libc functions, like syscall(), mmap(),
read(), write(), pread(), pwrite(), munmap(), open(), close(), with
wrapper functions which directly invoke the system call. This removes
most of libdius's dependency on libc. This makes the system call
implementation more efficent (since errno is no longer written to,
and the assembly can be directly inlined), and is the first step
to making libdius a run-time library which actually provides the
entire run-time.