di: implement binary search algorithms
di::binary_search() is a lot more usable than std::binary_search(),
because it returns a InFoundResult instead of a simple boolean. This
allows calling binary search to get an iterator to the found element,
and additionally, if the search fails, an iterator which points to
the insertion point. This is like Java's binary search, which returns
-1 - (insertion index) on failure, but in a much easier to decompose
way.

Additionally, since the comparators di uses don't just return a boolean,
and instead return a three way comparison result, checking for equality
does not require calling the comparator twice, and so the binary search
can early return as soon as the element is found.