di: implement intrusive ref-counting pointers
The underlying reference count machinery uses customization points
defined by both the pointer tag and the pointed to object. This lets
the underlying implementation choose whether or not to be thread safe,
which allows supporting Rc and Arc.

To use di::Arc, whatever object should be reference counted must
inherit from di::IntrusiveRefCount<Self>, which provides the object
a reference count stored in an atomic usize. For additional safety,
said type should have a private constructor and declare
di::IntrusiveRefCount<Self> as a friend. This ensures that the
objects cannot be created other than through the make_arc() and
try_make_arc() factory functions, which ensures that conversions
from Self to Arc<Self> will always be correct.

Reference counted types must not ever be stored by reference.
However, reference counted types should be passed by reference
when calling a function. If the function needs to store said
object, it can simply call object.arc_from_this(), or call the
di::AdoptObject overload constructor for Arc<Self>. Since the
reference count is intrusive, arc_from_this() can be called
inside an objects constructor, unlike std::shared_from_this().

The IntrusivePtr utilizes 2 type tags in its constructors and
modifying operations. AdoptObject signals that the constructor
should take a reference to the object, while RetainObject signals
the constructor does not need to take a reference. RetainObject is
useful when first creating objects, since their reference count is
initialized to 1. This is a low-level interface which can be used
when doing something custom. In practice, the object factories
make_arc() will take care of this for the user.

A downside of using IntrusivePtr's is that they are a lot less flexible
than std::shared_ptr. For instance, dynamic type casts and refererencing
subobjects cannot be supported, as well as having a pointer to builtin
or existing types. Additionally, we cannot easily have a Arc<T const>,
since the pointer itself needs to be mutable, in order to destroy T and
increment/decrement its reference count.

Although a lot less flexible, IntrusivePtr can use a much simpler implementation,
and only needs to store 1 pointer, not 2. Reference counted objects in the iris
kernel should aim to use Arc as much as possible.