iris: refactor code to store globals in iris::GlobalState
The makes the use of global variables in the kernel much more explicit.
For now, the global state is partial, since architecture specific
things and the physical page frame allocator are left out. In addition,
fields which need to be mutated are simply declared mutable rather than
actually being protected. But, by using this approach, when the kernel
actually needs to be thread safe, replacing mutable with di::Synchronized
will be a great start.