di: implement generic di::Any type erased container
di::Any is a templated class with 3 arguments, the erased interface,
the storage type, and the vtable policy.

The erased interface is provided by the user and signals what the
invocables the any should be callable for.

The storage type defines how any is constructed, and its corresponding
value semantics. A storage has an associated StorageCategory, which
controls which of the special member functions for Any are implemented,
and whether or not they are trivial.

The vtable policy defines the layout of the vtable. It is a policy because
depending on the storage type, the vtable may need to hold more entries
than the user specifically requested.

Generic aliases like AnyRef and AnyInline are provided which set default
values for the Storage and VTablePolicy template parameters.