di: implement StrongInt
This is a helper type which produces strongly typed integers. For
each unique type desired, the user simply defines a type tag with
the necessary information to create the wrapper. For flexibility,
StrongInt lets the Tag define its own mixin, which gives the wrapper
custom methods.

This will be used in the iris kernel to provide type-safe interfaces
for VirtualAddress and PhysicalAddress.