iris: refactor and improve shutdown mechanism
The Iris kernel now will automatically shutdown after panicking, which
is useful when running in an emulator. On real hardware, we would not
want to do this, so the user can inspect the error messages. But for
now, this is useful.

In addition, the kernel will attempt to communicate a shutdown status
code through the qemu isa-debug-exit device. This allows reporting
test failures to a test runner in a simple manner. A more complex kernel
could simply write the test results to some file system, or to a serial
port, but simplicity is needed at this stage in development.

The inspiration for using this mechanism comes from the rust os blog:
https://os.phil-opp.com/testing/