meta: enable running unit tests on Iros through ctest
This involves setting CMAKE_CROSSCOMPILING_EMULATOR to a custom
script which creates a kernel image and initrd, and then runs
qemu. The script reports success only if the emulator exits with
the magic status code 33, which is done using a magic Qemu device.

In combination with giving ctest a 30 second default timeout, this
actually works. Now, the tests can be run easily by just running
ctest. And soon, this will be done on CI!