iris: support x86_64 FPU state
This patch not only adds support for SSE2, but also AVX by supporting
using either xsave or fxsave. I don't have an AVX-512 CPU to test on, I
believe it would work without any code modification, since the FPU code
already asks the CPU for the size of the FPU state and enables all the
available bits in xcr0.