iris: implement serial input/output and run the shell!
Now, when running the Iris kernel, the user is provided a with a shell.
There are 4 programs available, `/test_create_task`, `/test_read`,
`/test_userspace`, and `/sh`. All paths provided must be absolute, the
shell has no prompt, and the shell exits if you make a typo. But, it is
still a functional shell, that can even spawn nested shells.