iris: destroy tasks on a dedicated finalizer task
Now, all tasks on the system are cleaned up by a dedicated finalizer
task. This is done because we need may need to free the task's memory
and kernel stack, and this cannot be done while executing in that task's
context. Additionally, the memory used to store task's kernel stacks is
now properly freed.