meta: setup kernel debugging via Qemu
This adds a new launch configuration which runs Qemu with GDB, and then
attaches to it. When building the kernel with debug symbols, this allows
for debugging the kernel like a normal executable. However, its is most
likely hard to use for problems which occur after the kernel has started
userspace.