iris: introduce and use Processor structure
This patch moves the formerly global scheduler into a Processor
structure. In addition, the current processor can be accessed using the
%gs register. This is properly managed as well to prevent kernel
addresses leaking to userspace. However, the code is written to assume
that the logical processor cannot change. This prevents code from being
scheduled on different processors if its running in the kernel and gets
pre-empted. In the future, code which accesses the current processor
should disable pre-emption explicitly.