di: introduce GuardedReference
This is a reference to a value that is protected by a RAII handle. This
is essentially the same class as LockedReference, but it works for any
RAII type. This can now be used in the kernel to protect references to
per-processor state, which must be protected by disabling preemption.