iris: protect access to processor state by disabling preemption
Any code which directly interacts with the current processor assumes
that this value will not change while it is running. This can only be
ensured by disabling preemption or interrupts. This patch makes use of
the newly added GuardedReference class to ensure that references to this
state are used correctly. An unsafe API is also provided for cases where
the caller knows that preemption/interrupts are disabled.

This is a very nice change, as it means that the kernel should safely
handle things once SMP is supported.