meta: update recommended cmake extensions
The new pre-release version of the CMake Tools extension has direct
support for CTest, which means the old extension for this is no longer
needed. Additionally, there is a newer extension which provides a CMake
language server, which produces very good completions. For best results,
users also have to disable the default CMake extension that ships with
CMake Tools.