iris: boot alternate processors
This commit adds very primitive support for booting the APs. However, if
there any APs, the system will crash. This is because the bootstrap
memory is not yet cleaned up properly. Additionally, the APs do not do
anything. On the bright side, getting the APs to boot into long mode is
really the hard part.