iris: implement TLB shootdown
This patch additionally refactors the booting of APs and driver
initialization to be after the scheduler is started. This allows
enabling interrupts on the BSP before the APs are started, which
prevents issues when the APs try to send IPIs to the BSP. Although
ultimately, the APs cannot allocate any memory until they start their
own schedulers, as interrupts need to be enabled to allocate memory
without deadlocking, given TLB shootdowns now happen.