iris: fix IRQ enabling in Scheduler::yield()
Instead of unconditionally enabling interrupts when returning from a
yield, force the caller to enable interrupts if they want to. This fixes
deadlock in cases where yield is called with interrupts disabled.

Additionally, the Scheduler::yield() function now is implemented more
efficently, by jumping directly the to return address instead of setting
a fixed continuation.