iris: fix spurious gsbase corruption
When a task made a system call, upon returning from the system call, the
kernel restored the gsbase value of the task, but did so without first
disabling interrupts. This meant that if an interrupt occurred while the
task was still in kernel-space, but had a user-space gsbase value, the
kernel would later read the corrupted gsbase value and triple fault when
trying to read the current processor.

The fix is to simply disable interrupts, and add an assertion which
checks that interrupts are disabled before touching the gsbase value.