iris: fix multi-processor scheduling
The main issue was that yielding corrupted the stack pointer, and that
IPI objects were not properly initialized. This commit fixes both of
those issues, allowing the kernel to work flawlessly on in SMP mode.

Since SMP is not enabled by default in the run script, the previous
broken commit is allowed to remain in the history, as a reminder to not
fail to initialize objects.