iris: implement fault aware copy_{to|from}_user
These make use of the fact that memcpy() can be implemented in a single
x86 instruction (rep movb). The kernel now installs a page fault handler
which will check if the faulting address this instruction. If it is, the
kernel will jump to a later point in the function, and return a non-zero
error code.

This commit also adds test cases, but currently there are several things
to fix. The first is preventing arbitrary read/writes from kernelspace,
the second is handling non-canonical addresses (which generate #GP
faults), and the kernel should also prevent creating userspace buffers
which have a combined base + size which overflows.