iris: use the APIC timer for scheduling
This commit replaces the PIT timer with the APIC timer for scheduling.
This is done by first using the PIT timer to calibrate the APIC timer,
and then setting up the APIC timer to fire an interrupt every 5ms. This
is done for each CPU, so SMP works even when KVM is disabled.