iris: use huge pages when mapping physical id map
It turns out that the most expensive part of the entire boot process was
setting up the physical id map (at least, without optimizations). Now,
instead of mapping millions of individual pages, the kernel can now make
use of 1 GiB mappings. In effect, for a 4 GiB address space, the kernel
now maps 4 pages instead of 1 million pages. Additionally, this requires
only a single physical page of overhead, instead of thousands of pages.

When 1 GiB mappings are not available, the kernel will fall back on
using 2 Mib mappings. This is still a significant improvement over the
previous implementation.