iris: tracking region backing objects for initial kernel regions
In particular, the kernel heap and executable are now tracked using
proper backing objects. This is a step towards being able to lazily
map pages in the kernel heap, and to use this mechanism for all memory
management.

Additionally, di::StrongInt is changed to support directly adding
unsigned types, which prevents a lot of casting. Unfortunately, this
introduces an ambiguity when using integer literals, but this can be
fixed by providing the correct type explicitly.