iris: use backing objects to free physical pages
Now, instead of blindly freeing all physical memory in an address space,
the kernel decrements the reference count of the individual pages. Each
region also has a reference counted object. This object also has owns a
reference to each page in the object. When mapping physical memory into
an address space, the page's reference count is incremented. It is
decremented when the page is unmapped.

This architecture will enable shared memory and memory mapped files, but
for now, there is a 1:1 mapping between physical pages and backing
objects.