meta: upgrade cross-compiler from GCC 12.2 to GCC 13.1
The gcc patches were modified to account for the fact that support for
the stabs debug format was removed.

Additionally, an initial patch for libstdc++ was added. This patch lets
the user try to compile libstdc++, but the build fails because the
libccpp is missing several functions, such as rand(), srand(), strtod(),
div(), sqort(), strpbrk(), etc.

Lastly, the CMake toolchain files were updated to compile position
independent code by default. This is necessary to link the system
libraries (which are statically linked for now), with gcc's shared
libraries. This change should probably be either part of a CMake
platform module, or be changed to the default in the compiler itself,
since position independent code is the default on Linux and is more
secure.