iris: implement and use basic path lookup
This commit replaces the previous adhoc initrd lookup with a proper VFS
abstraction. Currently, only the initrd is supported, but real file
systems can be supported in the future.

The path lookup itself does not yet support symbolic links, but does
support '..', '.', multiple slashes, and relative paths.

Unfortunately, this does break reading directory entries, since it is
not yet supported. Additionally, writing to inodes is not supported in
the VFS (which is not a change since the initrd never supported it).

Also, a temporary hack must be used to enable program loading from
files, since the implementation expects to be able to read the entire
inode in contiguous memory. This is possible for the initrd, but not any
other file system, since inode's store their content directly in the
page cache. This should be replaced by directly memorying mapping the
executable file into the process's address space on program load, or
making a much more complex ELF parser (but I think the first option is
better).

Obviously, additional concerns like permissions, append only files, non
regular files, memory mapping, asynchronicity, symlinks, hardlinks,
temporary inodes, and synchronization a not implemented.

Additionally, there is no system calls for getting/setting the current
working directory as well setting the task's root directory.