iris: implement tmpfs
This is a simple implementation of the tmpfs. It implements dynamic
inode creation, but does not support hard links or symlinks. The VFS
also doesn't support writing to files, but new files can be created.

In this commit, the mount mechanism is horribly broken, because mounts
do not show when listing their containing directory. This is because
the mount point is special cased in path lookup. The correct solution
is to attach the mount point directly to an existing directory, but this
will require changing the initrd generator to create the temp directory.