di: introduce functional chain() higher-order function
This function is like compose(f, g), but instead of creating a functor
which evaluates to f(g(x)), it creates a functor which evaluates to
g(f(x)). This allows writing functional code using a more imperitive
style, which fits the language better, and matches the usage of the '|'
operator in the ranges and execution libraries.

This commit also adds a dereference functor which just calls operator*,
and modifies reference wrapper's ref() and cref() functions to be
niebloids.