di: implement JSON value
This commit introduces a JSON value type, which is a variant of all
possible JSON types (object, string, etc.). This class additionally
provides member overloads for convienient access and mutation, so the
JSON value can be treated essentially as a JSON object or JSON array,
without the user having to down cast.

However, getting this JSON type to behave nicely was a challenge,
because it is a recursive type. To support this without indirection
the container classes di::Vector, di::TreeMap, and di::NodeHashMap were
changed to support incomplete types when declared. An additional problem
which arosed was defining operator== and operator<=> for JSON value,
since it resulted in circular concept constraints. This was solved by
changing these concepts to all their return value to be overridden by
the user, which the library does to effectively disable constraint
checking for these concepts when a JSON value is involved.