meta: use unity builds by default
This commit adds a CMakePreset generator that generates all the presets
needed, with both unity builds and non-unity builds. The unity builds
use the non-unity builds to generate the compile_commands.json file.
Additionally, this commit changes the build directories around, so that
is now reflected in the docs and relevant scripts.

Also, Iros tests should be retried once if they fail, which helps combat
flakiness. And the CI build now used builds with -Werror, which is not
the default locally.