di: make any container more like std::any
This replaces the `empty()` function with `has_value()`, and
additionally adds an `emplace()` function.

Additionally, when creating an any can fail, the provided create
function makes use of NRVO to avoid a potentially expesive move (since
for hybrid storage, this requires a vtable lookup). This also means that
immovable types can now be created.