iris: make reading a file async
This is a matter of changing the return type from Expected<usize> to
AnySenderOf<usize>. Additionally, the TRY_OR_SEND_ERROR macro is added
to make it easier to send errors when calling synchrnous functions from
an async context. The TRY_UNERASE_ERROR() macro is added to downcast the
error type to iris::Error, which is used when reporting system call
errors.

There are now a few places where di::execution::sync_wait() is called,
to block for the result of file reading. In the future, these callers
should be asynchronous themselves, which will allow these calls to be
removed.

As composing senders is a lot more difficult than using coroutines, it
would also be nice support using coroutines when return di::AnySender.
This is probably possible buts needs investigation.