di: add Stopped/stopped tag type
This is a tag type that can be used to indicate that a coroutine has
completed by stopping. The actual value of this tag is equivalent to the
return value di::execution::just_stopped(), which is for convenience.
This allows returning di::stopped from a function returning a
type-erased sender as well as from a coroutine.