di: begin implementation of zip for async sequences
This turns out to be extremely complicated, even after introducing
lockstep sequences. This commit only implements the outer logic, which
is effectively equivalent to the execution::when_all() algorithm. For
now, the returned sequence is just always empty.

Hopefully, the next commit will hook up the inner logic, which is going
to be a bit tricky.