meta: update doxygen to 1.9.7
This adds a new option which allows us to strip leading whitespace from
snippets, and this is now used to put the code examples referenced in
documention into the unit test file. This means that the code examples
will be tested, and we can be sure that they are correct.

For some reason, this new version changes the generated names of the
markdown pages, so we need to update the table of contents files to
refer to the new locations.

Lastly, the CI and Dockerfile have been updated to use the new version
of doxygen.