di: minor correctness fixups to ensure_started() and bulk()
ensure_started() should reset its stop callback when the future
completes.

ensure_started() should ASSERT() that the underlying shared state is not
the nullptr.

ensure_started() should allow the first completion signature to not be
default constructible, and instead use a dummy value as the first item
in the variant.

bulk() should make its operation state immovable, since the receiver
stores a pointer to it.