package_manager: mvp package building
This introduces the package_manager program, which will be used to
install ported software to Iros. Hopefully, this could replace the
adhoc build scripts for creating the toolchain as well. This program
is most likely similar to xbstrap in intended scope.

The MVP supports downloading packages, apply patches, and
building/installing with a Makefile. This general idea is to describe
each package using a JSON file, which describes its dependencies,
build system, and configuration instructions. The package manager
program reads these JSON files from the file system, and then attempts
to build the package passed on the command line. The JSON format is
described in a JSON schema file.