ports: add Figlet port
This ports the Figlet pretty-print cli. Unfortunately, the program
crashes the kernel when executing, for unknown reasons. A while ago,
the program was known to work. The issue could be an old version of
ccpp linked against the toolchain's libgcc, or there could be a more
fundamental issue with the kernel. Regardless, this demonstrates how
to use the package manager program.