di: introduce CommonError
This is a meta-programming facility like std::common_type, but for
erorrs. It is useful to combine errors from different sources, e.g. disk
IO and allocation failure.

The underlying mechanism defers to common type whenever possible, but
allows separate customizations, and natively supports di::Error, which
is a type-erased error. Additionally, the implementation falls back on
storing the errors as a variant if no common type is available.