di: implement sequence_to()
This function is equivalent to the haskell function sequence:

sequence :: T (M x) -> M (T x)

However, it only works for the di::Expected monad. It is used to reduce
a Con<Expected<T>> to an Expected<Con<T>>. However, it works for any
type of input and output container, and handles allocation failure.

The name comes from the fact that it is a mix between di::to() and
di::sequence(). The latter is essentially sequenceM_, just ignores the
inner values.