di: improve curry and curry_back functions
The interface is refactored so that functions can become curried by
simply inheriting from the Curry or CurryBack CRTP base class. They then
can provide a max_arity constant, and also must use the base class's
operator().

This is required for the erase and erase_if functions, which are both
customizable and curried.

Previously, the curry and curry_back functions were implemented as
wrapper functions, which resulted in a new function type being created,
and this broke the tag_invoke() based customization scheme.

These wrapper functions still exist, but they are now implemented in
terms of the new CRTP base classes. Additionally, the wrapper functions
return objects which can handle currying even if the wrapped function
excepts a variadic number of arguments. This is done by first trying to
call the wrapped function with the given arguments, and if that fails,
then the arguments are stored in a tuple, and a new function object is
returned. The key trick is that this new interface ensures that there is
never a case where multiple candidates are viable, and so the overload
resolution is always unambiguous.